Add plugin system
Currently inlined into golicons.js. Next step is to separate it
out. There's a bit of chicken and egg problem because
we need registerPlugin to exist before the plugins run, but the
plugins need to run before golicons starts messing with the DOM.
Plan is to create a separate init script to run after all plugins.